+++ src/structure/FlowNodeVisitor.ts
@@ -9,6 +9,7 @@ import { FlowIcon } from "../nodes/FlowIcon";
 import { FlowImage } from "../nodes/FlowImage";
 import { FlowNode } from "../nodes/FlowNode";
 import { FlowTable } from "../nodes/FlowTable";
+import { FlowVideo } from "../nodes/FlowVideo";
 import { InlineNode } from "../nodes/InlineNode";
 import { LineBreak } from "../nodes/LineBreak";
 import { ParagraphBreak } from "../nodes/ParagraphBreak";
@@ -91,6 +92,10 @@ export class FlowNodeVisitor implements GenericFlowNodeVisitor<FlowNode> {
     }
 
     visitImage(node: FlowImage): FlowNode {
+        return this.visitInline(node);
+    }
+
+    visitVideo(node: FlowVideo): FlowNode {
         return this.visitInline(node);
     }